# MAINT: count bfgs updates more carefully