This is an alternative to https://github.com/scipy/scipy/pull/4717, fixing a small oversight in https://github.com/scipy/scipy/pull/4572.