@argriffing 2015-04-20
Here's the starting point that caused the unanticipated `BFGS update SKIPPED` within a local refinement in the basinhopping search in https://github.com/scipy/scipy/pull/4717.

```python
from __future__ import print_function, division

import numpy as np
from scipy import optimize

def func2d(x):
    f = np.cos(14.5 * x[0] - 0.3) + (x[1] + 0.2) * x[1] + (x[0] + 0.2) * x[0]
    df = np.zeros(2)
    df[0] = -14.5 * np.sin(14.5 * x[0] - 0.3) + 2. * x[0] + 0.2
    df[1] = 2. * x[1] + 0.2
    return f, df

def main():
    troublesome_local_guess = np.array([0.01780837, -0.06329397])
    result = optimize.minimize(func2d, troublesome_local_guess,
            method='L-BFGS-B', jac=True, options={'disp' : True})

main()
```

```
RUNNING THE L-BFGS-B CODE

           * * *

Machine precision = 2.220D-16
 N =            2     M =           10
 This problem is unconstrained.

At X0         0 variables are exactly at the bounds

At iterate    0    f=  9.94354D-01    |proj g|=  8.41231D-01

At iterate    1    f= -9.58091D-02    |proj g|=  1.14293D+00
  ys=-3.043E-01  -gs= 9.035E-01 BFGS update SKIPPED

At iterate    2    f= -9.89999D-02    |proj g|=  1.11542D-01

At iterate    3    f= -9.92877D-02    |proj g|=  1.06100D-01

At iterate    4    f= -1.01633D-01    |proj g|=  3.05130D-01

At iterate    5    f= -1.02059D-01    |proj g|=  1.39710D-01

At iterate    6    f= -1.02110D-01    |proj g|=  1.51339D-02

At iterate    7    f= -1.02110D-01    |proj g|=  6.66463D-04

At iterate    8    f= -1.02110D-01    |proj g|=  2.32008D-05

At iterate    9    f= -1.02110D-01    |proj g|=  1.07182D-06

           * * *

Tit   = total number of iterations
Tnf   = total number of function evaluations
Tnint = total number of segments explored during Cauchy searches
Skip  = number of BFGS updates skipped
Nact  = number of active bounds at final generalized Cauchy point
Projg = norm of the final projected gradient
F     = final function value

           * * *

   N    Tit     Tnf  Tnint  Skip  Nact     Projg        F
    2      9     18      2     1     0   1.072D-06  -1.021D-01
  F = -0.10211034862284796     

CONVERGENCE: NORM_OF_PROJECTED_GRADIENT_<=_PGTOL            

 Cauchy                time 0.000E+00 seconds.
 Subspace minimization time 0.000E+00 seconds.
 Line search           time 0.000E+00 seconds.

 Total User time 0.000E+00 seconds.
```